+++ server/src/controllers/index.ts
@@ -1,0 +1,4 @@
+export { authController } from './authController';
+export { userController } from './userController';
+export { folderController } from './folderController';
+export { setController } from './setController';
