+++ server/src/entities/index.ts
@@ -1,0 +1,3 @@
+export { default as FolderEntity } from './FolderEntity';
+export { default as SetEntity } from './SetEntity';
+export { default as UserEntity } from './UserEntity';
