+++ server/src/lib/HttpException.ts
