+++ server/src/lib/index.ts
@@ -1,0 +1,2 @@
+export { HttpException } from './HttpException';
+export { logger } from './logger';
