+++ server/src/middlewares/errorMiddleware.ts
@@ -1,6 +1,6 @@
 import { NextFunction, Request, Response } from 'express';
-import { HttpException } from '@/utils/HttpException';
-import { logger } from '@utils/logger';
+import { HttpException } from '@src/lib/HttpException';
+import { logger } from '@src/lib/logger';
 
 const errorMiddleware = (error: HttpException, req: Request, res: Response, next: NextFunction) => {
   try {
