+++ server/src/middlewares/index.ts
@@ -1,0 +1,2 @@
+export { default as errorMiddleware } from './errorMiddleware';
+export { isAuth } from './isAuth';
