+++ server/src/routes/index.ts
@@ -1,0 +1,3 @@
+export { default as AuthRoute } from './authRoute';
+export { default as SetRoute } from './setRoute';
+export { default as FolderRoute } from './folderRoute';
