+++ server/src/services/index.ts
@@ -1,0 +1,4 @@
+export { authService } from './authService';
+export { folderService } from './folderService';
+export { setService } from './setService';
+export { userService } from './userService';
